project/backend/FASTAPI/requirements.txt
Package               Version
--------------------- -----------
annotated-doc         0.0.4
annotated-types       0.7.0
anyio                 4.12.0
argcomplete           1.10.3
beautifulsoup4        4.8.2
certifi               2025.11.12
chardet               3.0.4
charset-normalizer    3.4.4
click                 8.3.1
colorama              0.4.6
compressed-rtf        1.0.7
dnspython             2.8.0
docx2txt              0.9
ebcdic                1.1.1
email-validator       2.3.0
exceptiongroup        1.3.1
extract-msg           0.28.7
fastapi               0.128.0
fastapi-cli           0.0.20
fastapi-cloud-cli     0.8.0
fastar                0.8.0
filelock              3.20.1
fsspec                2025.12.0
h11                   0.16.0
httpcore              1.0.9
httptools             0.7.1
httpx                 0.28.1
huggingface-hub       0.36.0
idna                  3.11
IMAPClient            2.1.0
Jinja2                3.1.6
joblib                1.5.3
lxml                  6.0.2
markdown-it-py        4.0.0
MarkupSafe            3.0.3
mdurl                 0.1.2
mpmath                1.3.0
networkx              3.4.2
numpy                 2.2.6
olefile               0.47
packaging             25.0
pandas                2.3.3
pdfminer.six          20191110
pillow                12.0.0
pip                   22.2.1
pycryptodome          3.23.0
pydantic              2.12.5
pydantic_core         2.41.5
pydantic-extra-types  2.10.6
pydantic-settings     2.12.0
Pygments              2.19.2
python-dateutil       2.9.0.post0
python-dotenv         1.2.1
python-multipart      0.0.21
python-pptx           0.6.23
pytz                  2025.2
PyYAML                6.0.3
regex                 2025.11.3
requests              2.32.5
rich                  14.2.0
rich-toolkit          0.17.1
rignore               0.7.6
safetensors           0.7.0
scikit-learn          1.6.1
scipy                 1.15.3
sentence-transformers 5.2.0
sentry-sdk            2.48.0
setuptools            63.2.0
shellingham           1.5.4
six                   1.12.0
sortedcontainers      2.4.0
soupsieve             2.8.1
SpeechRecognition     3.8.1
starlette             0.50.0
sympy                 1.14.0
textract              1.6.5
threadpoolctl         3.6.0
tokenizers            0.22.1
tomli                 2.3.0
torch                 2.9.1
tqdm                  4.67.1
transformers          4.57.3
typer                 0.21.0
typing_extensions     4.15.0
typing-inspection     0.4.2
tzdata                2025.3
tzlocal               5.3.1
urllib3               2.6.2
uvicorn               0.40.0
watchfiles            1.1.1
websockets            15.0.1
xgboost               3.1.2
xlrd                  1.2.0
xlsxwriter            3.2.9
